[PERFORMANCE] 리액트 최적화 (#211)
* perf: 타이머 컴포넌트 분리 해서 랜더링 속도 증가

- timer 상태는 Timer 컴포넌트 안에서 관리

* chore: useMemo 커스텀훅을 useJobMemo로 이름 변경

* perf: 직업 메모 상태 관리를 MemoList 컴포넌트에서 수행하여 페이지 전체 리랜더링 방지

* refactor: Input 값 업데이트 함수 분리

인라인 함수를 따로 분리시켜 선언

* perf: 채팅 보내기 최적화

채팅을 입력해서 보내면 Game 페이지 전체가 리렌더링되는 문제 최적화

* chore: Timer파일 이름 index로 변경

* perf: 채팅 폼 컴포넌트 분리
